pyproject.toml
[tool.poetry]
name = "qa-guru-lesson-18"
version = "0.1.0"
description = ""
authors = ["Irina.Voron <yarutich.irina@gmail.com>"]
readme = "README.md"
packages = [{include = "qa"}]

[tool.poetry.dependencies]
python = "^3.12"
selene = "^2.0.0rc9"
allure-pytest = "^2.13.5"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
